--- .bookignore
@@ -1,7 +1,0 @@
-bin
-installer
-lib
-plugins
-scripts
-tasks
-test
