+++ docs/.vuepress/theme/styles/palette.styl
