+++ package.json
@@ -37,7 +37,7 @@
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "generate-tests": "leia \"examples/**/README.md\" test -r 6 -s 'Start up tests' -t 'Verification commands' -c 'Destroy tests' --split-file",
     "lint": "eslint --quiet --no-ignore bin lib plugins scripts test",
-    "lint:sites": "eslint  --quiet --ext .js,.vue --no-ignore --config .eslintrc.vuepress.json docs/.vuepress docs/.vuepress/theme/** website/.vuepress website/.vuepress/theme/**",
+    "lint:sites": "eslint  --quiet --ext .js,.vue --config .eslintrc.vuepress.json docs/.vuepress website/.vuepress",
     "pkg:cli": "node ./scripts/pkg.js",
     "pkg:full": "node ./scripts/pkg.js --installer",
     "release": "bump --prompt --tag --all --push",
