+++ website/.vuepress/theme/components/Home.vue
@@ -1,20 +1,5 @@
 <template>
-  <div class="theme-container" :class="pageClasses" @touchstart="onTouchStart" @touchend="onTouchEnd">
-    <Navbar v-if="shouldShowNavbar" @toggle-sidebar="toggleSidebar"/>
-
-    <div class="sidebar-mask" @click="toggleSidebar(false)"></div>
-
-    <Sidebar
-      :items="sidebarItems"
-      @toggle-sidebar="toggleSidebar">
-      <slot
-        name="sidebar-top"
-        slot="top"/>
-      <slot
-        name="sidebar-bottom"
-        slot="bottom"/>
-    </Sidebar>
-
+  <div>
     <div id="hero">
       <div class="inner">
         <div class="left">
@@ -148,32 +133,20 @@
     <div id="ready">
       <div class="inner">
         <h2>Ready for dev liberation?</h2>
-        <p><a class="button white" href="/download/">GET LANDO!</a></p>
+        <p><a class="button blue" href="/download/">GET LANDO!</a></p>
       </div>
     </div>
-
-    <MadeByTandem />
-    <Newsletter />
-    <Footer/>
   </div>
 </template>
 
 <script>
 // Core components and things
 import CarbonAds from '@theme/components/CarbonAds.vue';
-import Navbar from '@theme/components/Navbar.vue';
-import Sidebar from '@theme/components/Sidebar.vue';
-import {resolveSidebarItems} from '@parent-theme/util/index.js';
 
-// Lando components
-import MadeByTandem from '@theme/components/MadeByTandem';
-import Newsletter from '@theme/components/Newsletter';
-import Footer from '@theme/components/Footer';
 export default {
-  components: {CarbonAds, Sidebar, Navbar, MadeByTandem, Newsletter, Footer},
+  components: {CarbonAds},
   data() {
     return {
-      isSidebarOpen: false,
       data: {},
       growthRate: 0.0001,
       startingTime: 1565701419,
@@ -187,78 +160,9 @@ export default {
       const currentUsers = Math.floor(secondsSince * this.growthRate) + this.startingUsers;
       return currentUsers.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
     },
-    shouldShowNavbar() {
-      const {themeConfig} = this.$site;
-      const {frontmatter} = this.$page;
-      if (
-        frontmatter.navbar === false
-        || themeConfig.navbar === false) {
-        return false;
-      }
-      return (
-        this.$title
-        || themeConfig.logo
-        || themeConfig.repo
-        || themeConfig.nav
-        || this.$themeLocaleConfig.nav
-      );
-    },
-    shouldShowSidebar() {
-      const {frontmatter} = this.$page;
-      return (
-        !frontmatter.home
-        && frontmatter.sidebar !== false
-        && this.sidebarItems.length
-      );
-    },
-    sidebarItems() {
-      return resolveSidebarItems(
-        this.$page,
-        this.$page.regularPath,
-        this.$site,
-        this.$localePath
-      );
-    },
-    pageClasses() {
-      const userPageClass = this.$page.frontmatter.pageClass;
-      return [
-        {
-          'no-navbar': !this.shouldShowNavbar,
-          'sidebar-open': this.isSidebarOpen,
-          'no-sidebar': !this.shouldShowSidebar,
-        },
-        userPageClass,
-      ];
-    },
   },
   mounted() {
-    this.$router.afterEach(() => {
-      this.isSidebarOpen = false;
-    });
     this.data = this.$page.frontmatter;
-  },
-  methods: {
-    toggleSidebar(to) {
-      this.isSidebarOpen = typeof to === 'boolean' ? to : !this.isSidebarOpen;
-    },
-    // side swipe
-    onTouchStart(e) {
-      this.touchStart = {
-        x: e.changedTouches[0].clientX,
-        y: e.changedTouches[0].clientY,
-      };
-    },
-    onTouchEnd(e) {
-      const dx = e.changedTouches[0].clientX - this.touchStart.x;
-      const dy = e.changedTouches[0].clientY - this.touchStart.y;
-      if (Math.abs(dx) > Math.abs(dy) && Math.abs(dx) > 40) {
-        if (dx > 0 && this.touchStart.x <= 80) {
-          this.toggleSidebar(true);
-        } else {
-          this.toggleSidebar(false);
-        }
-      }
-    },
   },
 };
 </script>
@@ -484,16 +388,11 @@ export default {
   #carbon
     padding-bottom: 0
     background-color: darken($landoPink, 13%)
-  #news
-    background-color: darken($landoBlue, 12%)
-    padding: 2em 0
-    h3
-      font-size: 1.5em
-      a
-        color: lighten($landoGrey, 75%)
-  #made-by
-    padding: 7em 0
-    margin: 0
+  #ready
+    h2
+      color: lighten($landoPink, 80%)
+    p
+      font-style: normal
 @media (max-width: $MQNarrow)
   .lando-front
     #hero
