+++ website/.vuepress/theme/styles/index.styl
@@ -24,6 +24,10 @@ a.button,
       background-color: #4e6e8e
       color: $light
       border-color: #4e6e8e
+    &.blue
+      background-color: lighten($landoBlue, 25%)
+      color: #ffffff
+      border-color: lighten($landoBlue, 25%)
 .step
   padding-bottom: 2em
 .step-number
