+++ website/download/README.md
@@ -1,7 +1,5 @@
 ---
 navbar: true
-pageClass: lando-download
-layout: Download
 metaTitle: Download Lando
 description: Download Lando and take your local development and DevOps workflow to lightspeed.
 image: https://raw.githubusercontent.com/lando/lando/master/docs/.vuepress/public/images/hero-pink.png
