+++ website/memberships/README.md
@@ -1,7 +1,5 @@
 ---
 navbar: true
-pageClass: lando-memberships
-layout: Memberships
 metaTitle: The Lando Alliance
 description: Support Lando by joining The Alliance to liberate developers everywhere from the mind-forged manacles of unnecessary work, repeatable steps and dev monotony.
 image: https://raw.githubusercontent.com/lando/lando/master/docs/.vuepress/public/images/lando-alliance.png
